requirements.txt
streamlit>=1.34.0
openai>=0.28
python-dotenv>=0.21.0
